# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/internal/client-release-facts": {
      "post": {
        "operationId": "ClientReleaseFactsController_syncFacts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientReleaseFactsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientReleaseDetailDto"
                }
              }
            }
          }
        },
        "summary": "Sync client release facts from deploy executor",
        "tags": [
          "internal-client-releases"
        ]
      }
    },
    "/api/v1/admin/client-releases": {
      "get": {
        "operationId": "ClientReleasesAdminController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "example": "production",
              "type": "string"
            }
          },
          {
            "name": "client",
            "required": false,
            "in": "query",
            "schema": {
              "example": "appMobile",
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": false,
            "in": "query",
            "schema": {
              "example": "android",
              "type": "string"
            }
          },
          {
            "name": "distributionStatus",
            "required": false,
            "in": "query",
            "schema": {
              "example": "synced",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientReleaseListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List client shell releases",
        "tags": [
          "admin-client-releases"
        ]
      }
    },
    "/api/v1/admin/client-releases/packages": {
      "get": {
        "operationId": "ClientReleasesAdminController_listPackages",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "example": "production",
              "type": "string"
            }
          },
          {
            "name": "client",
            "required": false,
            "in": "query",
            "schema": {
              "example": "appMobile",
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": false,
            "in": "query",
            "schema": {
              "example": "android",
              "type": "string"
            }
          },
          {
            "name": "distributionStatus",
            "required": false,
            "in": "query",
            "schema": {
              "example": "synced",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientPackageListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List client release packages",
        "tags": [
          "admin-client-releases"
        ]
      }
    },
    "/api/v1/admin/client-releases/{id}": {
      "get": {
        "operationId": "ClientReleasesAdminController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientReleaseDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get client shell release detail",
        "tags": [
          "admin-client-releases"
        ]
      }
    },
    "/api/v1/admin/client-releases/{releaseId}/policies/{policyId}": {
      "patch": {
        "operationId": "ClientReleasesAdminController_updatePolicy",
        "parameters": [
          {
            "name": "releaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientReleasePolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientUpdatePolicySummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update client shell update policy",
        "tags": [
          "admin-client-releases"
        ]
      }
    },
    "/api/v1/client-downloads": {
      "get": {
        "operationId": "ClientDownloadsController_list",
        "parameters": [
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "example": "production",
              "type": "string"
            }
          },
          {
            "name": "client",
            "required": false,
            "in": "query",
            "schema": {
              "example": "appMobile",
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": false,
            "in": "query",
            "schema": {
              "example": "android",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDownloadInfoDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List latest available client downloads",
        "tags": [
          "client-downloads"
        ]
      }
    },
    "/api/v1/client-downloads/latest": {
      "get": {
        "operationId": "ClientDownloadsController_latest",
        "parameters": [
          {
            "name": "client",
            "required": true,
            "in": "query",
            "schema": {
              "example": "appMobile",
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": true,
            "in": "query",
            "schema": {
              "example": "android",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "example": "production",
              "type": "string"
            }
          },
          {
            "name": "currentVersion",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1.2.0",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDownloadInfoDto"
                }
              }
            }
          }
        },
        "summary": "Resolve latest client download URL",
        "tags": [
          "client-downloads"
        ]
      }
    },
    "/api/v1/client-updates/check": {
      "get": {
        "operationId": "ClientUpdatesController_check",
        "parameters": [
          {
            "name": "client",
            "required": true,
            "in": "query",
            "schema": {
              "example": "appMobile",
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": true,
            "in": "query",
            "schema": {
              "example": "android",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "example": "production",
              "type": "string"
            }
          },
          {
            "name": "currentVersion",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1.2.0",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDownloadInfoDto"
                }
              }
            }
          }
        },
        "summary": "Check client shell update availability",
        "tags": [
          "client-updates"
        ]
      }
    }
  },
  "info": {
    "title": "rtnn backend api",
    "description": "Template backend kernel for RTNN",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ClientReleaseFactsDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "example": "rtnn.deploy.client-release-facts.v1"
          },
          "environment": {
            "type": "string",
            "example": "production"
          },
          "generatedAt": {
            "type": "string",
            "example": "2026-04-30T06:24:12.000Z"
          },
          "mode": {
            "type": "string",
            "example": "write"
          },
          "project": {
            "type": "object"
          },
          "binding": {
            "type": "object"
          },
          "source": {
            "type": "object"
          },
          "release": {
            "type": "object"
          },
          "artifacts": {
            "type": "object"
          },
          "clients": {
            "type": "object"
          }
        },
        "required": [
          "schemaVersion",
          "environment",
          "clients"
        ]
      },
      "ClientPackageSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pkg_01JABCD123"
          },
          "client": {
            "type": "string",
            "example": "appMobile"
          },
          "target": {
            "type": "string",
            "example": "android"
          },
          "shell": {
            "type": "string",
            "example": "mobile"
          },
          "packageName": {
            "type": "string",
            "example": "com.example.app",
            "nullable": true
          },
          "artifactName": {
            "type": "string",
            "example": "app-release.apk"
          },
          "shellVersion": {
            "type": "string",
            "example": "1.2.3"
          },
          "releaseKind": {
            "type": "string",
            "example": "release"
          },
          "webUrl": {
            "type": "string",
            "example": "https://app.example.com",
            "nullable": true
          },
          "sourceUrl": {
            "type": "string",
            "example": "https://github.com/org/repo",
            "nullable": true
          },
          "distributionProvider": {
            "type": "string",
            "enum": [
              "github-release",
              "self-hosted-static",
              "external-url",
              "object-storage"
            ],
            "example": "github-release"
          },
          "distributionUrl": {
            "type": "string",
            "example": "https://download.example.com/app.apk",
            "nullable": true
          },
          "distributionStatus": {
            "type": "string",
            "enum": [
              "pending",
              "synced",
              "failed",
              "pruned",
              "disabled"
            ],
            "example": "synced"
          },
          "fileName": {
            "type": "string",
            "example": "app.apk",
            "nullable": true
          },
          "fileSize": {
            "type": "number",
            "example": 12345678,
            "nullable": true
          },
          "sha256": {
            "type": "string",
            "example": "abc123",
            "nullable": true
          },
          "signingStatus": {
            "type": "string",
            "example": "signed",
            "nullable": true
          },
          "buildStatus": {
            "type": "string",
            "example": "success",
            "nullable": true
          },
          "updaterStatus": {
            "type": "string",
            "example": "available",
            "nullable": true
          },
          "updaterUrl": {
            "type": "string",
            "example": "https://updates.example.com/latest.json",
            "nullable": true
          },
          "storeProvider": {
            "type": "string",
            "example": "google-play",
            "nullable": true
          },
          "storeStatus": {
            "type": "string",
            "example": "submitted",
            "nullable": true
          },
          "blockers": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "syncedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "prunedAt": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "client",
          "target",
          "shell",
          "artifactName",
          "shellVersion",
          "releaseKind",
          "distributionProvider",
          "distributionStatus",
          "blockers",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ClientUpdatePolicySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pol_01JABCD123"
          },
          "client": {
            "type": "string",
            "example": "appMobile"
          },
          "target": {
            "type": "string",
            "example": "android"
          },
          "channel": {
            "type": "string",
            "example": "production"
          },
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "recommendedReleaseId": {
            "type": "string",
            "example": "rel_01JABCD123",
            "nullable": true
          },
          "recommendedVersion": {
            "type": "string",
            "example": "1.2.3",
            "nullable": true
          },
          "releaseOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientUpdatePolicyReleaseOptionDto"
            }
          },
          "minimumSupportedVersion": {
            "type": "string",
            "example": "1.0.0",
            "nullable": true
          },
          "forceUpdate": {
            "type": "boolean",
            "example": false
          },
          "allowGithubFallback": {
            "type": "boolean",
            "example": true
          },
          "notes": {
            "type": "string",
            "example": "Roll out after store approval",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "client",
          "target",
          "channel",
          "enabled",
          "releaseOptions",
          "forceUpdate",
          "allowGithubFallback",
          "createdAt",
          "updatedAt"
        ]
      },
      "ClientReleaseDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "rel_01JABCD123"
          },
          "releaseVersion": {
            "type": "string",
            "example": "1.2.3"
          },
          "channel": {
            "type": "string",
            "example": "production"
          },
          "sourceRepository": {
            "type": "string",
            "example": "example/rtnn"
          },
          "sourceRunId": {
            "type": "string",
            "example": "123456789",
            "nullable": true
          },
          "sourceSha": {
            "type": "string",
            "example": "bce88fb0a427"
          },
          "sourceRef": {
            "type": "string",
            "example": "refs/heads/main",
            "nullable": true
          },
          "dryRun": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "example": "completed"
          },
          "generatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "syncedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "packageCount": {
            "type": "number",
            "example": 2
          },
          "downloadablePackageCount": {
            "type": "number",
            "example": 2
          },
          "clients": {
            "example": [
              "appMobile"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targets": {
            "example": [
              "android"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "distributionStatuses": {
            "example": [
              "synced"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "packages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientPackageSummaryDto"
            }
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientUpdatePolicySummaryDto"
            }
          }
        },
        "required": [
          "id",
          "releaseVersion",
          "channel",
          "sourceRepository",
          "sourceSha",
          "dryRun",
          "status",
          "packageCount",
          "downloadablePackageCount",
          "clients",
          "targets",
          "distributionStatuses",
          "createdAt",
          "updatedAt",
          "packages",
          "policies"
        ]
      },
      "ClientReleaseSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "rel_01JABCD123"
          },
          "releaseVersion": {
            "type": "string",
            "example": "1.2.3"
          },
          "channel": {
            "type": "string",
            "example": "production"
          },
          "sourceRepository": {
            "type": "string",
            "example": "example/rtnn"
          },
          "sourceRunId": {
            "type": "string",
            "example": "123456789",
            "nullable": true
          },
          "sourceSha": {
            "type": "string",
            "example": "bce88fb0a427"
          },
          "sourceRef": {
            "type": "string",
            "example": "refs/heads/main",
            "nullable": true
          },
          "dryRun": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "example": "completed"
          },
          "generatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "syncedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "packageCount": {
            "type": "number",
            "example": 2
          },
          "downloadablePackageCount": {
            "type": "number",
            "example": 2
          },
          "clients": {
            "example": [
              "appMobile"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targets": {
            "example": [
              "android"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "distributionStatuses": {
            "example": [
              "synced"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "releaseVersion",
          "channel",
          "sourceRepository",
          "sourceSha",
          "dryRun",
          "status",
          "packageCount",
          "downloadablePackageCount",
          "clients",
          "targets",
          "distributionStatuses",
          "createdAt",
          "updatedAt"
        ]
      },
      "ClientReleaseListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientReleaseSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "ClientPackageListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientPackageListItemDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateClientReleasePolicyDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "recommendedReleaseId": {
            "type": "string",
            "nullable": true
          },
          "minimumSupportedVersion": {
            "type": "string",
            "nullable": true,
            "example": "1.2.0"
          },
          "forceUpdate": {
            "type": "boolean"
          },
          "allowGithubFallback": {
            "type": "boolean"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ClientDownloadInfoDto": {
        "type": "object",
        "properties": {
          "client": {
            "type": "string",
            "example": "appMobile"
          },
          "target": {
            "type": "string",
            "example": "android"
          },
          "channel": {
            "type": "string",
            "example": "production"
          },
          "version": {
            "type": "string",
            "example": "1.2.3",
            "nullable": true
          },
          "shellVersion": {
            "type": "string",
            "example": "1.2.3",
            "nullable": true
          },
          "generatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "syncedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "nullable": true
          },
          "downloadType": {
            "type": "string",
            "enum": [
              "direct",
              "store",
              "unavailable"
            ],
            "example": "direct"
          },
          "provider": {
            "type": "string",
            "example": "github-release",
            "nullable": true
          },
          "downloadUrl": {
            "type": "string",
            "example": "https://download.example.com/app.apk",
            "nullable": true
          },
          "sourceUrl": {
            "type": "string",
            "example": "https://github.com/org/repo",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "example": "app.apk",
            "nullable": true
          },
          "fileSize": {
            "type": "number",
            "example": 12345678,
            "nullable": true
          },
          "sha256": {
            "type": "string",
            "example": "abc123",
            "nullable": true
          },
          "updateAvailable": {
            "type": "boolean",
            "example": true
          },
          "forceUpdate": {
            "type": "boolean",
            "example": false
          },
          "minimumSupportedVersion": {
            "type": "string",
            "example": "1.0.0",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "example": "Roll out after store approval",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "client",
          "target",
          "channel",
          "downloadType",
          "updateAvailable",
          "forceUpdate"
        ]
      }
    }
  }
}